Add files via upload
This is a dataset found on kaggle at https://www.kaggle.com/datasets/derrekdevon/real-estate-sales-2001-2020/data. It is about 97MB. Compressed, it was 28MB, which is too large to upload. Therefore, I  went into the dataset and deleted the bottom 350,000 or so rows, which reduced its file size to about 55MB. Compressed, it's now 17MB, which is small enough to load into a Github repository. I needed to upload this to my Github repository for peer graders to download and obtain the same results by running my R code in my final project.